fix shouldHideEvent design as per Kegan's comments
